python-service/requirements.txt
fastapi
uvicorn[standard]
spacy
nltk
textblob
pyinflect
language-tool-python
pdfplumber
python-docx
python-multipart
httpx
